Unify VolumeGeometry and VolumeVectorGeometry
Since a VolumeGeometry is effectively a VolumeVectorGeometry in a
single axis-aligned orientation, VolumeGeometry now wraps an inner
VolumeVectorGeometry ("has-a" rather than "is-a" relationship).

The two classes now both implement the following properties:
- pos, w, v, u
- num_steps
- shape
- size, sizes
- voxel_size, voxel_sizes
- voxel_volume, voxel_volumes
- corners

Most of the computation is performed in VolumeVectorGeometry.

In addition, VolumeGeometry implements `extent`, which is a popular
way to parametrize geometries in ASTRA and ODL.

+++ tomosipo/vector_calc.py
@@ -28,9 +28,9 @@ from contextlib import contextmanager
 
 
 def to_vec(x):
-    x = np.array(x, copy=False)
+    x = np.array(x, dtype=np.float64, copy=False)
     s = x.shape
-    x = np.array(x, copy=False, ndmin=2)
+    x = np.array(x, dtype=np.float64, copy=False, ndmin=2)
 
     if x.ndim == 2 and x.shape[1] == 3:
         return x
@@ -44,9 +44,9 @@ def to_vec(x):
 
 
 def to_scalar(x):
-    x = np.array(x, copy=False)
+    x = np.array(x, dtype=np.float64, copy=False)
     s = x.shape
-    x = np.array(x, copy=False, ndmin=1)
+    x = np.array(x, dtype=np.float64, copy=False, ndmin=1)
 
     if x.ndim == 1:
         # Add a single dimension to the right. Note that `to_vec` adds
